DataSubject and Engine structure (#1)
DataSubject is the parent class for all DataSource and DataTarget
objects.  Each DataSubject is defined by a Schema and an Engine.
The Engine is the mechanism by which data in a subject is represented
(would Interface be a better term?).

For example, a local file is a DataSource and a CSV file is an Engine.
The DataSource describes how to get the data (basically just a path to
the local file), and the Engine describes how to work with the data
(e.g., parse the CSV) according to the provided Schema.  Engines
should define mechanisms to convert to other engines.  For example, we
can convert data in a CSV engine to data in a Pandas engine.

This commit demonstrates the above example.