Working data target!
I revised the idea of an interface because I thought the
number of engines would explode and be difficult to manage
converting between them.

So, "engines" are now core data structures.  There should
be a smaller number of them (e.g., pandas, spark, elt) and
allowed conversions between them should be clearly identified.

There will be many more "interfaces".  Interfaces are the things
between source/targets and the engine.  Examples include
csv files, parquet files, dictionary of arrays, array of
dictionaries.  The interfaces are expected to know how
to convert to the core engines.

There will be even more data source and targets.
Data sources extract data from some place and load it into an
interface.  For example, an SFTP, local filesystem, or S3
could download csv files and pass those to a CSV interface.
Or, a REST API source could collect responses as an array
of dictionaries.  These results are then passed to a
supported interface.  Data targets work the same way,
but in reverse.

So a typical data pipeline then proceeds as:
DataSource
-> Interface
-> Engine
-> various transformations
-> Engine
-> Interface
-> DataTarget