Pipes!  Complete refactoring.
Got rid of the exceptionally confusing engine/interface/subject/pipeline
garbage.  Everything now is a pipe.  A pipe can take on any of the
roles that I was trying to use engine/interface/subject/pipeline to
do, but aren't constrained to that set alone.

Amazing what puting down a project for a few months can do.