Improve user experience when connecting pipes to self
Before this commit, a user could not create a new
pipe that had a source, and connect that source to the
source of nested pipe.  The workaround was to create
a target with the same name and connect the pipe's target
to the source of a nested pipe.  This was unnatural
and also prevented users from being able to also
connect the target of a nested pipe to the target
of the parent pipe.

This change modifies the way the Dask graph is built
to allow connecting the source of self to the source of
a nested pipe and the target of a nested pipe to the target
of self.